refactor: move docling_worker import to docling_utils for better organization
